requirements.txt
alembic==1.12.1
annotated-types==0.6.0
anyio==3.7.1
blinker==1.7.0
certifi==2023.7.22
click==8.1.7
colorama==0.4.6
Deprecated==1.2.14
distro==1.8.0
Flask==3.0.0
Flask-Limiter==3.5.0
Flask-Migrate==4.0.5
Flask-SQLAlchemy==3.1.1
greenlet==3.0.1
h11==0.14.0
httpcore==1.0.2
httpx==0.25.1
idna==3.4
importlib-resources==6.1.1
itsdangerous==2.1.2
Jinja2==3.1.2
limits==3.6.0
Mako==1.3.0
markdown-it-py==3.0.0
MarkupSafe==2.1.3
mdurl==0.1.2
openai==1.2.4
ordered-set==4.1.0
packaging==23.2
pydantic==2.5.0
pydantic_core==2.14.1
Pygments==2.16.1
python-dotenv==1.0.0
rich==13.6.0
sniffio==1.3.0
SQLAlchemy==2.0.23
tqdm==4.66.1
typing_extensions==4.8.0
Werkzeug==3.0.1
wrapt==1.16.0
